+++ tests/Models/TestModelNoPermission.php
@@ -7,7 +7,7 @@ use Pulsar\Model;
 
 class TestModelNoPermission extends ACLModel
 {
-    protected function hasPermission($permission, Model $requester)
+    protected function hasPermission($permission, Model $requester): bool
     {
         return false;
     }